DB utc changed to bigint, minor backend

+++ src/db/mysql.sql
@@ -3,7 +3,7 @@ CREATE DATABASE message_app;
 USE message_app;
 
 CREATE TABLE messages (
-	id int AUTO_INCREMENT NOT NULL,
+	id bigint AUTO_INCREMENT NOT NULL,
     content text NOT NULL,
     createdUtc int NOT NULL,
     primary key (id)
